LayoutTests/http/tests/security/xss-DENIED-assign-location-nonstandardProperty-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/xss-DENIED-assign-location-nonstandardProperty.html from frame with URL http://localhost:8000/security/resources/xss-DENIED-assign-location-nonstandardProperty-attacker.html. Domains, protocols and ports must match.


PASS: cross-site assignment of location.nonstandardProperty not allowed
